+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## Version 0.0.1 (2020-02-26)
 ### Added
 - Initial release.
+
+## Version 0.0.2 (2020-06-02)
+### Bug
+- Log users and teams that the script user does not have permissions to.
+
+## Version 0.0.3 (2020-06-03)
+### Added
+- user-teams report: Ability to only report teams that have specific members.